Crate transcribestreaming_code_example

Macros

stream

Asynchronous stream

Structs

AudioEvent

Provides a wrapper for the audio chunks that you are sending.

Blob

Binary Blob Type

Client

An ergonomic service client for Transcribe.

Duration

A Duration type to represent a span of time, typically used for system timeouts.

Opt

Region

The region to send requests to.

RegionProviderChain

Load a region by selecting the first from a series of region providers.

Enums

AudioStream

Represents the audio stream from your application to Amazon Transcribe.

Error

All possible error types for this service.

LanguageCode

MediaEncoding

TranscriptResultStream

Represents the transcription result stream from Amazon Transcribe to your application.

Constants

CHUNK_SIZE

Statics

PKG_VERSION

Crate version number.

Traits

BufMut

A trait for values that provide sequential write access to bytes.

StructOpt

A struct that is converted from command line arguments.

Functions

main

Transcribes an audio file to text.

pcm_data

Derive Macros

StructOpt

Generates the StructOpt impl.